requirements.txt
asgiref>=3.8.1
Django==5.2.6
django-ranged-response>=0.2.0
django-recaptcha>=4.1.0
django-simple-captcha==0.6.2
et_xmlfile>=2.0.0
mysqlclient>=2.2.7
openpyxl>=3.1.5
pillow>=11.2.1
python-dotenv>=1.1.0
sqlparse>=0.5.3
typing_extensions>=4.13.2
tzdata>=2025.2